refactor: cars, count 멤버들을 관리하는 Manage 클래스를 만들어 관심사를 분리
